Move screens to their own file in their own functions
