+++ src/index.js
@@ -3,9 +3,9 @@ require('colors');
 const { spawn } = require('child_process');
 const fs = require('fs');
 const path = require('path');
-const Table = require('cli-table');
 const defaultSettings = require('./default-settings.js');
 const ErrorWithoutStack = require('./error-without-stack.js');
+const screens = require('./screens.js');
 const utils = require('./utils.js');
 
 
@@ -46,20 +46,8 @@ module.exports = function Constructor(customSettings) {
 	
 	// Handle --version
 	if (settings.arguments.includes('-v') || settings.arguments.includes('--version')) {
-		// Print name, if we have one
-		if (settings.app.name) {
-			process.stdout.write((`${settings.app.name}: `).bold);
-		}
-		
-		
-		// Print version
-		console.log(`${settings.app.version}`);
-		
-		
-		// Add spacing after
-		for (let i = 0; i < settings.spacing.after; i += 1) {
-			console.log();
-		}
+		// Output
+		process.stdout.write(screens(settings).version());
 		
 		
 		// Verbose output
@@ -73,214 +61,8 @@ module.exports = function Constructor(customSettings) {
 	
 	// Handle --help
 	if (settings.arguments.length === 0 || settings.arguments.includes('-h') || settings.arguments.includes('--help')) {
-		// Get all commands in this program
-		let commands = utils(settings).getAllProgramCommands();
-		
-		
-		// Verbose output
-		utils(settings).verboseLog(`Found commands: ${commands.join(' | ')}`);
-		
-		
-		// Add leading space to commands
-		commands = commands.map(command => ` ${command}`);
-		const preppedCommand = ` ${organizedArguments.command}`.replace(/^ $/g, '');
-		
-		
-		// Filter out any unnecessary commands, which are...
-		commands = commands.filter(command => command.substring(0, preppedCommand.length) === `${preppedCommand}`); // Not related to current command
-		commands = commands.filter(command => (command.match(/ /g) || []).length === ((preppedCommand.match(/ /g) || []).length + 1)); // Not a direct sub-command of the current command
-		
-		
-		// Strip current command off front
-		commands = commands.map(command => command.substring(preppedCommand.length + 1));
-		
-		
-		// Sort commands
-		commands.sort();
-		
-		
-		// Verbose output
-		utils(settings).verboseLog(`Processable: ${commands.join(' | ')}`);
-		
-		
-		// Get merged spec for this command
-		const mergedSpec = utils(settings).getMergedSpec(organizedArguments.command);
-		
-		
-		// Determine if there are flags
-		let hasFlags = false;
-		
-		if (Object.entries(mergedSpec.flags).length) {
-			hasFlags = true;
-		}
-		
-		
-		// Determine if there are options
-		let hasOptions = false;
-		
-		if (Object.entries(mergedSpec.options).length) {
-			hasOptions = true;
-		}
-		
-		
-		// Determine if it accepts data
-		let acceptsData = false;
-		
-		if (mergedSpec.data && mergedSpec.data.description) {
-			acceptsData = true;
-		}
-		
-		
-		// Determine if there are commands
-		let hasCommands = false;
-		
-		if (commands.length) {
-			hasCommands = true;
-		}
-		
-		
-		// Output description
-		if (mergedSpec.description) {
-			console.log(`${'Description:'.bold} ${mergedSpec.description}`);
-		}
-		
-		
-		// Form and output usage line
-		let usageLine = `${'Usage:'.bold} ${settings.usageCommand}${(organizedArguments.command ? ` ${organizedArguments.command}` : '')}`;
-		usageLine += `${(hasCommands ? ' [commands]' : '').gray}`;
-		usageLine += `${(hasFlags ? ' [flags]' : '').gray}`;
-		usageLine += `${(hasOptions ? ' [options]' : '').gray}`;
-		usageLine += `${(acceptsData ? ' [data]' : '').gray}`;
-		
-		console.log(usageLine);
-		
-		
-		// Initialize table
-		const table = new Table({
-			chars: {
-				top: '',
-				'top-mid': '',
-				'top-left': '',
-				'top-right': '',
-				bottom: '',
-				'bottom-mid': '',
-				'bottom-left': '',
-				'bottom-right': '',
-				left: '',
-				'left-mid': '',
-				mid: '',
-				'mid-mid': '',
-				right: '',
-				'right-mid': '',
-				middle: '    ',
-			},
-			style: {
-				'padding-left': 0,
-				'padding-right': 0,
-			},
-		});
-		
-		
-		// Handle flags
-		if (hasFlags) {
-			// Add header
-			table.push(['\nFLAGS:']);
-			
-			
-			// List flags
-			Object.entries(mergedSpec.flags).forEach(([flag, details]) => {
-				table.push([`  --${flag}${details.shorthand ? `, -${details.shorthand}` : ''}`, `${details.description ? `${details.description}` : ''}`]);
-			});
-		}
-		
-		
-		// Handle options
-		if (hasOptions) {
-			// Add header
-			table.push(['\nOPTIONS:']);
-			
-			
-			// List options
-			Object.entries(mergedSpec.options).forEach(([option, details]) => {
-				// Form full description
-				let fullDescription = '';
-				
-				if (details.description) {
-					fullDescription += details.description;
-				}
-				
-				if (details.required) {
-					fullDescription += ' (required)'.gray.italic;
-				}
-				
-				if (details.type) {
-					fullDescription += ` (${details.type})`.gray.italic;
-				}
-				
-				if (details.accepts) {
-					fullDescription += ` (accepts: ${details.accepts.join(', ')})`.gray.italic;
-				}
-				
-				
-				// Add to table
-				table.push([`  --${option}${details.shorthand ? `, -${details.shorthand}` : ''}`, fullDescription]);
-			});
-		}
-		
-		
-		// Add line for commands
-		if (hasCommands) {
-			// Add header
-			table.push(['\nCOMMANDS:']);
-		}
-		
-		
-		// Loop over and push each command
-		commands.forEach((command) => {
-			const mergedSpec = utils(settings).getMergedSpec(`${organizedArguments.command} ${command}`.trim());
-			table.push([`  ${command}`, mergedSpec.description ? mergedSpec.description : '']);
-		});
-		
-		
-		// Print table
-		console.log(table.toString());
-		
-		
-		// Handle data
-		if (acceptsData) {
-			// Print header
-			console.log('\nDATA:');
-			
-			
-			// Form full description
-			let fullDescription = '';
-			
-			if (mergedSpec.data.description) {
-				fullDescription += mergedSpec.data.description;
-			}
-			
-			if (mergedSpec.data.required) {
-				fullDescription += ' (required)'.gray.italic;
-			}
-			
-			if (mergedSpec.data.type) {
-				fullDescription += ` (${mergedSpec.data.type})`.gray.italic;
-			}
-			
-			if (mergedSpec.data.accepts) {
-				fullDescription += ` (accepts: ${mergedSpec.data.accepts.join(', ')})`.gray.italic;
-			}
-			
-			
-			// Print
-			console.log(`  ${fullDescription}`);
-		}
-		
-		
-		// Add spacing after
-		for (let i = 0; i < settings.spacing.after; i += 1) {
-			console.log();
-		}
+		// Output
+		process.stdout.write(screens(settings).help());
 		
 		
 		// Verbose output
